+++ ext/java/nokogiri/XmlSaxParserContext.java
@@ -1,10 +1,14 @@
 package nokogiri;
 
 import nokogiri.internals.*;
+import static nokogiri.internals.NokogiriHelpers.rubyStringToString;
+
 import org.apache.xerces.parsers.AbstractSAXParser;
 import org.jruby.Ruby;
 import org.jruby.RubyClass;
+import org.jruby.RubyEncoding;
 import org.jruby.RubyFixnum;
+import org.jruby.RubyString;
 import org.jruby.anno.JRubyClass;
 import org.jruby.anno.JRubyMethod;
 import org.jruby.exceptions.RaiseException;
@@ -14,6 +18,7 @@ import org.jruby.runtime.builtin.IRubyObject;
 import org.xml.sax.SAXException;
 import org.xml.sax.SAXParseException;
 
+import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 
@@ -90,16 +95,26 @@ public class XmlSaxParserContext extends ParserContext
    * Create a new parser context that will parse the string
    * <code>data</code>.
    */
-  @JRubyMethod(name = "memory", meta = true)
+  @JRubyMethod(name = "native_memory", meta = true)
   public static IRubyObject
-  parse_memory(ThreadContext context,
-               IRubyObject klazz,
-               IRubyObject data)
+  parse_memory(ThreadContext context, IRubyObject klazz, IRubyObject data, IRubyObject encoding)
   {
-    final Ruby runtime = context.runtime;
-    XmlSaxParserContext ctx = newInstance(runtime, (RubyClass) klazz);
-    ctx.initialize(runtime);
-    ctx.setStringInputSource(context, data, runtime.getNil());
+    String java_encoding = null;
+    if (encoding != context.runtime.getNil()) {
+      if (!(encoding instanceof RubyEncoding)) {
+        throw context.runtime.newTypeError("encoding must be kind_of Encoding");
+      }
+      java_encoding = ((RubyEncoding)encoding).toString();
+    }
+
+    XmlSaxParserContext ctx = newInstance(context.runtime, (RubyClass) klazz);
+    ctx.initialize(context.runtime);
+    ctx.setStringInputSourceNoEnc(context, data, context.runtime.getNil());
+
+    if (java_encoding != null) {
+      ctx.getInputSource().setEncoding(java_encoding);
+    }
+
     return ctx;
   }
 
@@ -107,16 +122,26 @@ public class XmlSaxParserContext extends ParserContext
    * Create a new parser context that will read from the file
    * <code>data</code> and parse.
    */
-  @JRubyMethod(name = "file", meta = true)
+  @JRubyMethod(name = "native_file", meta = true)
   public static IRubyObject
-  parse_file(ThreadContext context,
-             IRubyObject klazz,
-             IRubyObject data)
+  parse_file(ThreadContext context, IRubyObject klazz, IRubyObject data, IRubyObject encoding)
   {
-    final Ruby runtime = context.runtime;
-    XmlSaxParserContext ctx = newInstance(runtime, (RubyClass) klazz);
-    ctx.initialize(context.getRuntime());
+    String java_encoding = null;
+    if (encoding != context.runtime.getNil()) {
+      if (!(encoding instanceof RubyEncoding)) {
+        throw context.runtime.newTypeError("encoding must be kind_of Encoding");
+      }
+      java_encoding = ((RubyEncoding)encoding).toString();
+    }
+
+    XmlSaxParserContext ctx = newInstance(context.runtime, (RubyClass) klazz);
+    ctx.initialize(context.runtime);
     ctx.setInputSourceFile(context, data);
+
+    if (java_encoding != null) {
+      ctx.getInputSource().setEncoding(java_encoding);
+    }
+
     return ctx;
   }
 
@@ -126,21 +151,30 @@ public class XmlSaxParserContext extends ParserContext
    *
    * TODO: Currently ignores encoding <code>enc</code>.
    */
-  @JRubyMethod(name = "io", meta = true)
+  @JRubyMethod(name = "native_io", meta = true)
   public static IRubyObject
-  parse_io(ThreadContext context,
-           IRubyObject klazz,
-           IRubyObject data,
-           IRubyObject encoding)
+  parse_io(ThreadContext context, IRubyObject klazz, IRubyObject data, IRubyObject encoding)
   {
-    // check the type of the unused encoding to match behavior of CRuby
-    if (!(encoding instanceof RubyFixnum)) {
-      throw context.getRuntime().newTypeError("encoding must be kind_of String");
+    if (!invoke(context, data, "respond_to?", context.runtime.newSymbol("read")).isTrue()) {
+      throw context.runtime.newTypeError("argument expected to respond to :read");
     }
-    final Ruby runtime = context.runtime;
-    XmlSaxParserContext ctx = newInstance(runtime, (RubyClass) klazz);
-    ctx.initialize(runtime);
-    ctx.setIOInputSource(context, data, runtime.getNil());
+
+    String java_encoding = null;
+    if (encoding != context.runtime.getNil()) {
+      if (!(encoding instanceof RubyEncoding)) {
+        throw context.runtime.newTypeError("encoding must be kind_of Encoding");
+      }
+      java_encoding = ((RubyEncoding)encoding).toString();
+    }
+
+    XmlSaxParserContext ctx = newInstance(context.runtime, (RubyClass) klazz);
+    ctx.initialize(context.runtime);
+    ctx.setIOInputSource(context, data, context.runtime.getNil());
+
+    if (java_encoding != null) {
+      ctx.getInputSource().setEncoding(java_encoding);
+    }
+
     return ctx;
   }