+++ ext/nokogiri/html4_sax_parser_context.c
@@ -2,52 +2,56 @@
 
 VALUE cNokogiriHtml4SaxParserContext ;
 
+/* :nodoc: */
 static VALUE
-noko_html4_sax_parser_s_parse_memory(VALUE klass, VALUE data, VALUE encoding)
+noko_html4_sax_parser_context_s_native_memory(VALUE rb_class, VALUE rb_input, VALUE rb_encoding)
 {
-  htmlParserCtxtPtr ctxt;
-
-  Check_Type(data, T_STRING);
-
-  if (!(int)RSTRING_LEN(data)) {
+  Check_Type(rb_input, T_STRING);
+  if (!(int)RSTRING_LEN(rb_input)) {
     rb_raise(rb_eRuntimeError, "input string cannot be empty");
   }
 
-  ctxt = htmlCreateMemoryParserCtxt(StringValuePtr(data),
-                                    (int)RSTRING_LEN(data));
-  if (ctxt->sax) {
-    xmlFree(ctxt->sax);
-    ctxt->sax = NULL;
+  if (!NIL_P(rb_encoding) && !rb_obj_is_kind_of(rb_encoding, rb_cEncoding)) {
+    rb_raise(rb_eTypeError, "argument must be an Encoding object");
   }
 
-  if (RTEST(encoding)) {
-    xmlCharEncodingHandlerPtr enc = xmlFindCharEncodingHandler(StringValueCStr(encoding));
-    if (enc != NULL) {
-      xmlSwitchToEncoding(ctxt, enc);
-      if (ctxt->errNo == XML_ERR_UNSUPPORTED_ENCODING) {
-        rb_raise(rb_eRuntimeError, "Unsupported encoding %s",
-                 StringValueCStr(encoding));
-      }
-    }
+  htmlParserCtxtPtr c_context =
+    htmlCreateMemoryParserCtxt(StringValuePtr(rb_input), (int)RSTRING_LEN(rb_input));
+  if (!c_context) {
+    rb_raise(rb_eRuntimeError, "failed to create xml sax parser context");
   }
 
-  return noko_xml_sax_parser_context_wrap(klass, ctxt);
+  noko_xml_sax_parser_context_set_encoding(c_context, rb_encoding);
+
+  if (c_context->sax) {
+    xmlFree(c_context->sax);
+    c_context->sax = NULL;
+  }
+
+  return noko_xml_sax_parser_context_wrap(rb_class, c_context);
 }
 
+/* :nodoc: */
 static VALUE
-noko_html4_sax_parser_context_s_parse_file(VALUE klass, VALUE filename, VALUE encoding)
+noko_html4_sax_parser_context_s_native_file(VALUE rb_class, VALUE rb_filename, VALUE rb_encoding)
 {
-  htmlParserCtxtPtr ctxt = htmlCreateFileParserCtxt(
-                             StringValueCStr(filename),
-                             StringValueCStr(encoding)
-                           );
-
-  if (ctxt->sax) {
-    xmlFree(ctxt->sax);
-    ctxt->sax = NULL;
+  if (!NIL_P(rb_encoding) && !rb_obj_is_kind_of(rb_encoding, rb_cEncoding)) {
+    rb_raise(rb_eTypeError, "argument must be an Encoding object");
   }
 
-  return noko_xml_sax_parser_context_wrap(klass, ctxt);
+  htmlParserCtxtPtr c_context = htmlCreateFileParserCtxt(StringValueCStr(rb_filename), NULL);
+  if (!c_context) {
+    rb_raise(rb_eRuntimeError, "failed to create xml sax parser context");
+  }
+
+  noko_xml_sax_parser_context_set_encoding(c_context, rb_encoding);
+
+  if (c_context->sax) {
+    xmlFree(c_context->sax);
+    c_context->sax = NULL;
+  }
+
+  return noko_xml_sax_parser_context_wrap(rb_class, c_context);
 }
 
 static VALUE
@@ -84,10 +88,10 @@ noko_init_html_sax_parser_context(void)
   cNokogiriHtml4SaxParserContext = rb_define_class_under(mNokogiriHtml4Sax, "ParserContext",
                                    cNokogiriXmlSaxParserContext);
 
-  rb_define_singleton_method(cNokogiriHtml4SaxParserContext, "memory",
-                             noko_html4_sax_parser_s_parse_memory, 2);
-  rb_define_singleton_method(cNokogiriHtml4SaxParserContext, "file",
-                             noko_html4_sax_parser_context_s_parse_file, 2);
+  rb_define_singleton_method(cNokogiriHtml4SaxParserContext, "native_memory",
+                             noko_html4_sax_parser_context_s_native_memory, 2);
+  rb_define_singleton_method(cNokogiriHtml4SaxParserContext, "native_file",
+                             noko_html4_sax_parser_context_s_native_file, 2);
 
   rb_define_method(cNokogiriHtml4SaxParserContext, "parse_with",
                    noko_html4_sax_parser_context__parse_with, 1);