+++ test/html4/sax/test_parser_context.rb
@@ -33,6 +33,104 @@ module Nokogiri::HTML4::SAX
         assert_raises(TypeError) { ParserContext.io(0xcafecafe) }
         assert_raises(TypeError) { ParserContext.file(0xcafecafe) }
       end
+
+      describe "encoding" do
+        # this input string is really ISO-8859-1 but is marked as UTF-8
+        let(:html_encoding_broken2) { <<~HTML }
+          <html><meta charset="UTF-8">
+          <body>B\xF6hnhardt</body>
+        HTML
+
+        it "gracefully handles nonsense encodings" do
+          assert_raises(ArgumentError) do
+            ParserContext.io(StringIO.new("asdf"), "not-an-encoding")
+          end
+          assert_raises(ArgumentError) do
+            ParserContext.memory("asdf", "not-an-encoding")
+          end
+          assert_raises(ArgumentError) do
+            ParserContext.file(Nokogiri::TestCase::XML_FILE, "not-an-encoding")
+          end
+        end
+
+        describe ".io" do
+          it "supports passing encoding name" do
+            pc = ParserContext.io(StringIO.new(html_encoding_broken2), "ISO-8859-1")
+            parser = Parser.new(Nokogiri::SAX::TestCase::Doc.new)
+            pc.parse_with(parser)
+
+            assert_equal("Böhnhardt", parser.document.data.join.strip)
+          end
+
+          it "supports passing Encoding" do
+            pc = ParserContext.io(StringIO.new(html_encoding_broken2), Encoding::ISO_8859_1)
+            parser = Parser.new(Nokogiri::SAX::TestCase::Doc.new)
+            pc.parse_with(parser)
+
+            assert_equal("Böhnhardt", parser.document.data.join.strip)
+          end
+
+          it "supports passing libxml2 encoding id" do
+            enc = nil
+            assert_output(nil, /deprecated/) do
+              enc = Parser::ENCODINGS["ISO-8859-1"]
+            end
+
+            pc = nil
+            assert_output(nil, /deprecated/) do
+              pc = ParserContext.io(StringIO.new(html_encoding_broken2), enc)
+            end
+
+            parser = Parser.new(Nokogiri::SAX::TestCase::Doc.new)
+            pc.parse_with(parser)
+
+            assert_equal("Böhnhardt", parser.document.data.join.strip)
+          end
+        end
+
+        describe ".memory" do
+          it "supports passing encoding name" do
+            pc = ParserContext.memory(html_encoding_broken2, "ISO-8859-1")
+            parser = Parser.new(Nokogiri::SAX::TestCase::Doc.new)
+            pc.parse_with(parser)
+
+            assert_equal("Böhnhardt", parser.document.data.join.strip)
+          end
+
+          it "supports passing Encoding" do
+            pc = ParserContext.memory(html_encoding_broken2, Encoding::ISO_8859_1)
+            parser = Parser.new(Nokogiri::SAX::TestCase::Doc.new)
+            pc.parse_with(parser)
+
+            assert_equal("Böhnhardt", parser.document.data.join.strip)
+          end
+        end
+
+        describe ".file" do
+          let(:file) do
+            Tempfile.new.tap do |f|
+              f.write html_encoding_broken2
+              f.close
+            end
+          end
+
+          it "supports passing encoding name" do
+            pc = ParserContext.file(file.path, "ISO-8859-1")
+            parser = Parser.new(Nokogiri::SAX::TestCase::Doc.new)
+            pc.parse_with(parser)
+
+            assert_equal("Böhnhardt", parser.document.data.join.strip)
+          end
+
+          it "supports passing Encoding" do
+            pc = ParserContext.file(file.path, Encoding::ISO_8859_1)
+            parser = Parser.new(Nokogiri::SAX::TestCase::Doc.new)
+            pc.parse_with(parser)
+
+            assert_equal("Böhnhardt", parser.document.data.join.strip)
+          end
+        end
+      end
     end
 
     describe "#parse_with" do